added registration.tsx and separate the login and register form
